+++ packages/yt/lib/src/oauth/oauth_access_control.dart
@@ -1,6 +1,6 @@
-import 'package:googleapis_auth/googleapis_auth.dart';
+import 'package:oauth2/oauth2.dart' as oauth2;
 
 import 'oauth_access_control_interface.dart';
 
-OAuthAccessControl getOAuthAccessControl(ClientId? clientId) =>
+OAuthAccessControl getOAuthAccessControl(oauth2.Client? oauthClient) =>
     throw UnimplementedError('getOAuthAccessControl');